requirements2.txt
numpy==1.18.1
streamlit==0.82.0
av==8.0.3
opencv_python_headless==4.5.2.54
pandas==0.25.3
matplotlib==3.1.2
Pillow==8.2.0
streamlit_webrtc==0.20.1
